requirements.txt
-i https://pypi.org/simple
blinker==1.7.0 ; python_version >= '3.8'
click==8.1.7 ; python_version >= '3.7'
flask==3.0.0
importlib-metadata==6.8.0 ; python_version < '3.10'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
markupsafe==2.1.3 ; python_version >= '3.7'
python-decouple==3.8
werkzeug==3.0.1 ; python_version >= '3.8'
zipp==3.17.0 ; python_version >= '3.8'
